step4 - refactor CarCollection 일급 컬렉션 변경
